[ncl] Rewrite a screen with geocoding examples
