+++ apps/native-component-list/src/navigation/ExpoApis.ts
@@ -31,7 +31,6 @@ const FaceDetector = optionalRequire(() => require('../screens/FaceDetectorScree
 const FileSystem = optionalRequire(() => require('../screens/FileSystemScreen'));
 const FirebaseRecaptcha = optionalRequire(() => require('../screens/FirebaseRecaptchaScreen'));
 const Font = optionalRequire(() => require('../screens/FontScreen'));
-const Geocoding = optionalRequire(() => require('../screens/GeocodingScreen'));
 const Google = optionalRequire(() => require('../screens/GoogleScreen'));
 const GoogleSignIn = optionalRequire(() => require('../screens/GoogleSignInScreen'));
 const Haptics = optionalRequire(() => require('../screens/HapticsScreen'));
@@ -98,7 +97,6 @@ const optionalScreens: {
   Contacts,
   ContactDetail,
   Events,
-  Geocoding,
   ImageManipulator,
   ImagePicker,
   InAppPurchases,
