+++ apps/native-component-list/src/screens/Location/LocationScreens.ts
@@ -1,9 +1,11 @@
 import BackgroundLocationMapScreen from './BackgroundLocationMapScreen';
 import GeofencingScreen from './GeofencingScreen';
 import LocationScreen from './LocationScreen';
+import GeocodingScreen from './GeocodingScreen';
 
 export default {
   Location: LocationScreen,
+  Geocoding: GeocodingScreen,
   Geofencing: GeofencingScreen,
   BackgroundLocationMap: BackgroundLocationMapScreen,
 };
